refactor: move auth-activity-provider to contexts directory
